requirements.txt
anndata==0.10.3
cellxgene-census
openpyxl
dataclasses_json
pandas
ruamel.yaml
jsonschema==4.4.0
ordered-set==4.1.0
deepmerge==1.1.0
cell-annotation-schema